+++ assets/js/base/components/cart-checkout/index.js
@@ -1,4 +1,5 @@
 export { default as AddressForm } from './address-form';
+export { default as Button } from './button';
 export { default as CheckoutForm } from './form';
 export { default as FormStep } from './form-step';
 export { default as PlaceOrderButton } from './place-order-button';
