+++ assets/js/base/components/cart-checkout/shipping-calculator/address.js
@@ -3,8 +3,10 @@
  */
 import PropTypes from 'prop-types';
 import { __ } from '@wordpress/i18n';
-import { AddressForm } from '@woocommerce/base-components/cart-checkout';
-import Button from '@woocommerce/base-components/button';
+import {
+	AddressForm,
+	Button,
+} from '@woocommerce/base-components/cart-checkout';
 import { useState } from '@wordpress/element';
 import isShallowEqual from '@wordpress/is-shallow-equal';
 import { useValidationContext } from '@woocommerce/base-context';
